(return codes) - Parse function return codes.
With this commit, the Parse function will return an integer indicating
the outcome of the parse operation.

Before this commit, the Parse function used to return a boolean value.
And, in case of a --help or --version flag, the function would display
the appropriate information.

Now, the function will return specific values. Also, it's up to the
caller to display the help message by executing the new ShowDisplay
function.

New ShowVersion function. Version 0.3.0.