Improving action and flag validation
Expecting errors in flag and action validation. Removing the
ValidateActionArguments() since the action validation will be done
during the parsing, as for the flag validation.
Replacing PARSE_ tokens with the new ParseResult enum; the Parse()
function now returns a ParseResult value instead of an int.

Updating example and unit tests.
Updating README.