Fix Action flag aliasing
Here we fix an issue where looking up an action flag by an aliased name
returned the wrong result.

This was due to us creating a new flag object for every action flag in
the action context, which meant that (unlike other flags) -f and --foo
pointed to different objects.