Add variable_arity flag to actions
We previously used to set the number of expected arguments of an action
by passing negative values for the arity as an indication of
number of arguments. A consequence of that was that we could not
register an action with 0 or more arguments.

With this commit we use a specific flag for indicating that a variable
number of arguments will be passed, whereas the arity value will
indicate the minmum number of arguments.