Terminate arg list of unit tests with a nullptr
Unit tests for the Parse() function should include a nullptr at the end
of the argument list in order to replicate the behaviour of the argv
list provided by the OS at runtime.

This will make sure that whenever HorseWhisperer tries to go out the
boundary of the argument list, a EXC_BAD_ACCESS is emitted. In that way,
we can effectively rely on the tests.