Return the action exit code in case of failure
With this commit the Start() function will return 0 if all requested
actions succeeded, 1 in case the callback of a requested action was not
defined, or the exit code of a failed action.

Note that HorseWhisperer will consider action as failed in case it does
not return 0.

The previous behaviour was to return always 1 in case an action failed,
regardless of its exit code.